requirements.txt
pandas
yfinance
mysql-connector-python
matplotlib
streamlit
